Ouroboros.Network.Protocol.LocalStateQuery.Type

Description

The type of the local ledger state query protocol.

This is used by local clients (like wallets and CLI tools) to query the ledger state of a local node.

Documentation

data LocalStateQuery block point (query :: Type -> Type) where

The kind of the local state query protocol, and the types of the states in the protocol state machine.

It is parametrised over the type of block (for points), the type of queries and query results.

Constructors

StIdle :: LocalStateQuery block point query
The client has agency. It can ask to acquire a state or terminate.

There is no timeout in this state.

StAcquiring :: LocalStateQuery block point query
The server has agency. it must acquire the state at the requested point or report a failure.

There is a timeout in this state.

StAcquired :: LocalStateQuery block point query
The client has agency. It can request queries against the current state, or it can release the state.

StQuerying :: result -> LocalStateQuery block point query
The server has agency. It must respond with the query result.

StDone :: LocalStateQuery block point query
Nobody has agency. The terminal state.

data Target point

Constructors

VolatileTip
The tip of the volatile chain

Cannot fail to be acquired.

SpecificPoint point
A specified point

Fails to be acquired if the point is not between VolatileTip and ImmutableTip (inclusive).

ImmutableTip
The tip of the immutable chain

Cannot fail to be acquired.

Requires at least NodeToClientV_16.

data AcquireFailure

Constructors

AcquireFailurePointTooOld

AcquireFailurePointNotOnChain

class (forall result. Show (query result)) => ShowQuery query where

To implement Show for:

('Message' ('LocalStateQuery' block query) st st')

we need a way to print the query GADT and its type index, result. This class contain the method we need to provide this Show instance.

We use a type class for this, as this Show constraint propagates to a lot of places.

Methods

showResult :: forall result. query result -> result -> String