Ouroboros.Network.Protocol.PeerSharing.Type

Documentation

newtype PeerSharingAmount

PeerSharing amount new type.

We use Word8 to be faithful to the CDDL specification.

Constructors

PeerSharingAmount

Fields

getAmount :: Word8

data PeerSharingResult peerAddress

PeerSharing Result type.

We need a constructor for the case when the Governor wins the race versus the Mux (when initialising the peer sharing miniprotocol). This leads the Governor to lookup a peer that hasn't been registered yet.

Constructors

PeerSharingResult [peerAddress]

PeerSharingNotRegisteredYet

data PeerSharing peerAddress where

A kind to identify our protocol, and the types of the states in the state transition diagram of the protocol.

Constructors

StIdle :: PeerSharing peerAddress
The client can send a request and the server is waiting for a request.

StBusy :: PeerSharing peerAddress
The server is responsible for sending response back.

StDone :: PeerSharing peerAddress
Both the client and server are in the terminal state. They're done.