Ouroboros.Network.Protocol.PeerSharing.Codec

Documentation

codecPeerSharing ∷ ∀ m peerAddress. MonadST m ⇒ (peerAddress → Encoding) → (∀ s. Decoder s peerAddress) → Codec (PeerSharing peerAddress) DeserialiseFailure m ByteString

codecPeerSharingId ∷ ∀ peerAddress m. Monad m ⇒ Codec (PeerSharing peerAddress) CodecFailure m (AnyMessage (PeerSharing peerAddress))

maxTransmissionUnit ∷ Word

We assume that a TCP segment size of 1440 bytes with initial window of size 4. This sets upper limit of 5760 bytes on each message of peer sharing protocol, which means request and response should be done in a single RTT

byteLimitsPeerSharing ∷ (bytes → Word) → ProtocolSizeLimits (PeerSharing peerAddress) bytes

timeLimitsPeerSharing ∷ ProtocolTimeLimits (PeerSharing peerAddress)